refactor: move mock methods to `test_utils`
