+++ crates/block-producer/src/test_utils/mod.rs
@@ -5,7 +5,7 @@ use tokio::sync::RwLock;
 
 mod proven_tx;
 
-pub use proven_tx::MockProvenTxBuilder;
+pub use proven_tx::{mock_proven_tx, MockProvenTxBuilder};
 
 mod store;
 
@@ -18,3 +18,5 @@ pub use account::MockPrivateAccount;
 pub mod block;
 
 pub mod batch;
+
+pub mod note;
