+++ crates/utils/src/lib.rs
@@ -2,4 +2,3 @@ pub mod config;
 pub mod errors;
 pub mod formatting;
 pub mod logging;
-pub mod testing;
